+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+1.0.0-rc.44
+===========
+* Added DELETE HTTP Request method
+* Added an ability upload/download files
+
 1.0.0-rc.43
 ===========
 * Added an ability scroll to up if element is overlapped by another element